chore: use tanvity as fallback ai search
