+++ frontend/rust-lib/flowy-search-pub/src/lib.rs
@@ -1,2 +1,6 @@
 pub mod cloud;
 pub mod entities;
+pub mod tantivy_state;
+pub mod tantivy_state_init;
+
+mod schema;
