+++ frontend/rust-lib/flowy-search/src/lib.rs
@@ -3,5 +3,4 @@ pub mod entities;
 pub mod event_handler;
 pub mod event_map;
 pub mod protobuf;
-mod schema;
 pub mod services;
