+++ frontend/rust-lib/flowy-server/src/local_server/impls/search.rs
@@ -5,17 +5,61 @@ use flowy_ai_pub::cloud::search_dto::{
 };
 use flowy_error::FlowyError;
 use flowy_search_pub::cloud::SearchCloudService;
+use flowy_search_pub::entities::TanvitySearchResponseItem;
+use flowy_search_pub::tantivy_state::DocumentTantivyState;
+use flowy_server_pub::search_dto::SearchContentType;
 use lib_infra::async_trait::async_trait;
-use std::sync::Arc;
+use std::sync::{Arc, Weak};
+use tokio::sync::RwLock;
+use tracing::trace;
 use uuid::Uuid;
 
 pub struct LocalSearchServiceImpl {
   #[allow(dead_code)]
   pub logged_user: Arc<dyn LoggedUser>,
   pub local_ai: Arc<LocalAIController>,
+  pub state: Option<Weak<RwLock<DocumentTantivyState>>>,
 }
 
-impl LocalSearchServiceImpl {}
+impl LocalSearchServiceImpl {
+  async fn local_search(
+    &self,
+    workspace_id: &Uuid,
+    query: &str,
+  ) -> Option<Vec<SearchDocumentResponseItem>> {
+    match self.state.as_ref().and_then(|v| v.upgrade()) {
+      None => {
+        trace!("[Search] tanvity state is None");
+        None
+      },
+      Some(state) => {
+        let results = state.read().await.search(workspace_id, query, None).ok()?;
+        let items = results
+          .into_iter()
+          .flat_map(|v| tanvity_document_to_search_document(*workspace_id, v))
+          .collect::<Vec<_>>();
+        Some(items)
+      },
+    }
+  }
+}
+
+fn tanvity_document_to_search_document(
+  workspace_id: Uuid,
+  doc: TanvitySearchResponseItem,
+) -> Option<SearchDocumentResponseItem> {
+  let object_id = Uuid::parse_str(&doc.id).ok()?;
+  Some(SearchDocumentResponseItem {
+    object_id,
+    workspace_id,
+    score: 1.0,
+    content_type: Some(SearchContentType::PlainText),
+    content: doc.content,
+    preview: None,
+    created_by: "".to_string(),
+    created_at: Default::default(),
+  })
+}
 
 #[async_trait]
 impl SearchCloudService for LocalSearchServiceImpl {
@@ -28,11 +72,23 @@ impl SearchCloudService for LocalSearchServiceImpl {
     {
       if let Ok(scheduler) = flowy_ai::embeddings::context::EmbedContext::shared().get_scheduler() {
         match scheduler.search(workspace_id, &query).await {
-          Ok(results) => return Ok(results),
-          Err(err) => tracing::error!("Local AI search failed: {:?}", err),
+          Ok(results) => {
+            return if results.is_empty() {
+              trace!("[Search] Local AI search returned no results, falling back to local search");
+              let items = self
+                .local_search(workspace_id, &query)
+                .await
+                .unwrap_or_default();
+              trace!("[Search] Local search returned {} results", items.len());
+              Ok(items)
+            } else {
+              Ok(results)
+            }
+          },
+          Err(err) => tracing::error!("[Search] Local AI search failed: {:?}", err),
         }
       } else {
-        tracing::error!("Could not acquire local AI scheduler");
+        tracing::error!("[Search] Could not acquire local AI scheduler");
       }
     }
 
@@ -47,6 +103,11 @@ impl SearchCloudService for LocalSearchServiceImpl {
   ) -> Result<SearchSummaryResult, FlowyError> {
     #[cfg(any(target_os = "windows", target_os = "macos", target_os = "linux"))]
     {
+      if search_results.is_empty() {
+        trace!("[Search] No search results to summarize");
+        return Ok(SearchSummaryResult { summaries: vec![] });
+      }
+
       if let Ok(scheduler) = flowy_ai::embeddings::context::EmbedContext::shared().get_scheduler() {
         let setting = self.local_ai.get_local_ai_setting();
         match scheduler
@@ -60,6 +121,8 @@ impl SearchCloudService for LocalSearchServiceImpl {
         tracing::error!("Could not acquire local AI scheduler");
       }
     }
+
+    //
     Ok(SearchSummaryResult { summaries: vec![] })
   }
 }
